Improvements: gzip under flag, remove temp user agent header and better file structure
