+++ mashrl/Cargo.toml
@@ -12,5 +12,9 @@ repository = "https://github.com/kaleidawave/experiments"
 [lib]
 path = "lib.rs"
 
+[features]
+decompress = ["dep:flate2"]
+
 [dependencies]
+flate2 = { version = "1.1.2", optional = true }
 native-tls = "0.2.14"
