+++ mashrl/examples/get.rs
@@ -1,8 +1,3 @@
-Minimal and simple HTTP(s) request library
-
-Usage:
-
-```rust
 use mashrl::{HTTP, make_get_request};
 use std::io::Read;
 
@@ -19,18 +14,3 @@ fn main() {
     response.body.read_to_string(&mut content).unwrap();
     eprintln!("{content}");
 }
-```
-
-Features
-
-- Simple to use API
-- Header creation and parsing API
-- Chunked decoding
-- Gzip decompression (under `-F decompress`)
-
-Future
-
-- Specify port (default to `443`, the standard for *internet traffic*)
-- Improvements to body handling (auto `Content-Length` header?)
-- Tracing (under flag)?
-- HTTPS under flag?
